work on sql stuff + added LoggedAction enum
Made it so that the project uses the bukkit javadoc url for bukkit
classes
Work on the JukeAlertLogger , mainly the the schema for the tables to be
more sane (why were we storing the location of the snitch in the
snitches detail table? we should just be storing the id of the row in
teh main snitchesTable and use that for the location, etc). Also work on
the getSnitchInfo method, using the updated schema and a new
PreparedStatement to get the snitch_id of a snitch from teh table based
on a Location, should probably put that into a method..
Also created the LoggedAction enum class, which will be used to insert /
restore actions inside the snitchesDetailTable, so we are not storing
strings like "block_break" which is error prone and takes up more space
then we need.