Feature Addition:  Juke Lever Triggering
Included is new code to allow new feature "Juke Lever Triggering" to
exist.  Specifically, this code will allow jukeboxes (not noteblocks)
that have opted in to lever triggering to cause a lever to momentarily
fire (~0.75 second ish) such that cardinal north when a player
enters/logs in a snitch radius, cardinal south when a player opens a
chest, cardinal east when a player places a block, and cardinal west
when a player breaks a block.  This new feature is safeguarded by new
configuration setting (default false) "allowTriggeringLevers".  If
false, which it is by default, then there should be no server-affect of
this change except a new column added to the main snitches table.  If
true, then it will allow users (owners/moderators of a snitch) to call
new command /jaToggleLevers that will allow them to opt-in a juke box
into this new functionality.  It is possible to remove this
functionality after it has gone live by modifying the configuration
option with no harm caused.  While levers were chosen for this
functionality, it could easily have been buttons as they offer the same
benefit (redstone pulse) and are part of the bukkit supported mechanism
for said functionality - levers were kept simply because the client
visible affect looks more interesting and novel.   Churned plugin.yml
version.